+++ src/main/java/memethespire/shoppatches/ShopRelicPurchasePatch.java
@@ -1,12 +1,10 @@
-package memethespire.cardmemes;
+package memethespire.shoppatches;
 
 import com.evacipated.cardcrawl.modthespire.lib.SpirePatch;
 import com.evacipated.cardcrawl.modthespire.lib.SpirePostfixPatch;
 import com.megacrit.cardcrawl.shop.ShopScreen;
 import com.megacrit.cardcrawl.shop.StoreRelic;
 import memethespire.ReflectionUtils;
-
-import java.lang.reflect.Field;
 
 @SpirePatch(clz = StoreRelic.class, method = "purchaseRelic")
 public class ShopRelicPurchasePatch {
@@ -16,6 +14,6 @@ public class ShopRelicPurchasePatch {
     public static void updateCards(StoreRelic instance) {
         ShopScreen shopScreen = (ShopScreen) ReflectionUtils.getPrivate(
                 instance, StoreRelic.class, "shopScreen");
-        ShopInitCardsPatch.modifyCards(shopScreen);
+        ShopInitCardsPatch.modifyCardsAndRelics(shopScreen);
     }
 }