+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 #### Features
 
 * Added cards contain check in `PlayerConditions` through the `cardsContain` attribute in the conditions field of meme entries.
+* Added a new type of meme: relic modification, where the names and descriptions of relics can be modified based on the player's current condition.
 
 #### Changes